resolve cihub/seelog#110 : explode archived logs.
**Why ?**
Popular command line utilities like `zcat` or `zgrep` are useful tools for processing log files. For instance, `zgrep` allow to search compressed files for a regular expression.
Those tools require that the analyzed archive contains only one file. Yet, when compressing archived logs, seelog put all the log files in the same archive.

So, The purpose of this pull request is to add the possibility for archived logs to be zipped in dedicated zip files.

**How ?**
As new compression formats might be introduced in future release of `seelog`, we want to clearly separate the choice of the compression algorithm to the choice of the gathering mode (grouped vs exploded).
Therefore, we propose a new attribute, `archiveexploded` (a boolean), specifying if the logs should be exploded or grouped inside the same archive file.