Add SyslogPriority formatter
Add a new formatter for the syslog priority field. This produces a token
"<PRIORITY>" with an integer produced by bitwise OR of syslog facility
(LOG_DAEMON, LOG_USER etc.) and priority (LOG_INFO, LOG_ERR etc.).

A sensible default facility and seelog-level-to-priority mapping is
included, so "%SyslogPriority" can be used straight away with no further
configuration/code in formats. The DAEMON facility is chosen as Go is
often used to write servers/daemons. However, it is possible to
customise both the facility and the level mapping if desired.