Add NewSyslogWriter and <syslog> receiver
Adds a <syslog> receiver that will write messages to a local syslog
daemon. This element defaults to using the std:syslog-bsd-local format
but can be overridden in the configuration file.

If writing to a remote syslog daemon (over the network) is needed, then
simply use the conn writer with an appropriate format (std:syslog-bsd or
std:syslog-rfc5424) instead.

NewSyslogWriter() simply returns a connWriter that has been configured
for use with a local syslog. It uses the detection logic from Go's
standard library to find the Unix socket it should connect to.