Rolling log file writer: avoid truncating file when rotating
There are several conditions under which the rotating log file writer
can actually truncate the file it is rotating.

Principally, this is caused by a lack of locking when writing to the log
file. In rollingFileWriter.Write(), we test for whether the current file
needs to be rotated, and if so then perform the rotation. This needs
locking to ensure that we do not have two (or more) goroutines making
the decision to rotate, then rotating. If this occurs then the second
rotation of an empty or virtually-empty file will truncate the original
file's contents. This commit resolves the problem by adding mutex
locking to Write().

There is also a more convoluted path through the time-based file writer
code which (previously) checked the modification time of the file on
disk, meaning that a mutex in the Write() routine alone is not
sufficient.  This commit also changes the time-based tests to cache the
time and filename, removing the need for a call to os.Stat() and
removing the possibility of making an inconsistent decision to roll more
than once.

Fixes #143.